fix: Move component to component file
